task 13
only delete states with 'a'

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -18,7 +18,8 @@ if __name__ == "__main__":
     session = Session()
 
     for state in session.query(State).all():
-        session.delete(state)
+        if "a" in state.name:
+            session.delete(state)
 
     session.commit()
